+++ client/src/components/EventView/Invitee/index.js
